Add interface blocks
