+++ assets/course-theme/blocks/ui/index.js
@@ -1,0 +1,7 @@
+/**
+ * Internal dependencies
+ */
+import spacerBlock from './spacer-block';
+import uiBlock from './ui-block.js';
+
+export default [ uiBlock, spacerBlock ];
